Rearrange protractor tests so there is one test in each suite.
