+++ core/tests/protractor/learnerDashboard.js
@@ -13,22 +13,26 @@
 // limitations under the License.
 
 /**
- * @fileoverview End-to-end tests for the learner dashboard page.
+ * @fileoverview End-to-end tests for the learner dashboard page and
+ * subscriptions functionality.
  */
+
+var editor = require('../protractor_utils/editor.js');
+var general = require('../protractor_utils/general.js');
+var users = require('../protractor_utils/users.js');
+var workflow = require('../protractor_utils/workflow.js');
 
 var AdminPage = require('../protractor_utils/AdminPage.js');
 var CreatorDashboardPage =
   require('../protractor_utils/CreatorDashboardPage.js');
 var CollectionEditorPage =
   require('../protractor_utils/CollectionEditorPage.js');
-var editor = require('../protractor_utils/editor.js');
-var general = require('../protractor_utils/general.js');
-var LibraryPage = require('../protractor_utils/LibraryPage.js');
 var ExplorationPlayerPage =
   require('../protractor_utils/ExplorationPlayerPage.js');
-var users = require('../protractor_utils/users.js');
 var LearnerDashboardPage =
   require('../protractor_utils/LearnerDashboardPage.js');
+var LibraryPage = require('../protractor_utils/LibraryPage.js');
+var PreferencesPage = require('../protractor_utils/PreferencesPage.js');
 var SubscriptionDashboardPage =
   require('../protractor_utils/SubscriptionDashboardPage.js');
 
@@ -243,6 +247,87 @@ describe('Learner dashboard functionality', function() {
     learnerDashboardPage.expectFeedbackExplorationTitleToMatch('About Oppia');
     learnerDashboardPage.navigateToFeedbackThread();
     learnerDashboardPage.expectFeedbackMessageToMatch(feedback);
+    users.logout();
+  });
+
+  afterEach(function() {
+    general.checkForConsoleErrors([]);
+  });
+});
+
+
+describe('Subscriptions functionality', function() {
+  var creatorDashboardPage = null;
+  var preferencesPage = null;
+  var subscriptionDashboardPage = null;
+
+  beforeEach(function() {
+    creatorDashboardPage = new CreatorDashboardPage.CreatorDashboardPage();
+    preferencesPage = new PreferencesPage.PreferencesPage();
+    subscriptionDashboardPage = (
+      new SubscriptionDashboardPage.SubscriptionDashboardPage());
+  });
+
+  it('handle subscriptions to creators correctly', function() {
+    // Create two creators.
+    users.createUser('creator1@subscriptions.com', 'creator1subscriptions');
+    users.login('creator1@subscriptions.com');
+    workflow.createExploration();
+    general.waitForSystem();
+    general.waitForSystem();
+    users.logout();
+
+    users.createUser('creator2@subscriptions.com', 'creator2subscriptions');
+    users.login('creator2@subscriptions.com');
+    workflow.createExploration();
+    users.logout();
+
+    // Create a learner who subscribes to both the creators.
+    users.createUser('learner1@subscriptions.com', 'learner1subscriptions');
+    users.login('learner1@subscriptions.com');
+    browser.get('/profile/creator1subscriptions');
+    browser.waitForAngular();
+    subscriptionDashboardPage.navigateToSubscriptionButton();
+    browser.get('/profile/creator2subscriptions');
+    browser.waitForAngular();
+    subscriptionDashboardPage.navigateToSubscriptionButton();
+    preferencesPage.get();
+    preferencesPage.expectDisplayedFirstSubscriptionToBe('creator...');
+    preferencesPage.expectDisplayedLastSubscriptionToBe('creator...');
+    users.logout();
+
+    // Create a learner who subscribes to one creator and unsubscribes from the
+    // other.
+    users.createUser('learner2@subscriptions.com', 'learner2subscriptions');
+    users.login('learner2@subscriptions.com');
+    browser.get('/profile/creator1subscriptions');
+    browser.waitForAngular();
+    subscriptionDashboardPage.navigateToSubscriptionButton();
+    browser.get('/profile/creator2subscriptions');
+    browser.waitForAngular();
+    // Subscribe and then unsubscribe from the same user.
+    subscriptionDashboardPage.navigateToSubscriptionButton();
+    browser.waitForAngular();
+    subscriptionDashboardPage.navigateToSubscriptionButton();
+    preferencesPage.get();
+    preferencesPage.expectSubscriptionCountToEqual(1);
+    preferencesPage.expectDisplayedFirstSubscriptionToBe('creator...');
+    users.logout();
+
+    users.login('creator1@subscriptions.com');
+    creatorDashboardPage.get();
+    browser.waitForAngular();
+    creatorDashboardPage.navigateToSubscriptionDashboard();
+    subscriptionDashboardPage.expectSubscriptionFirstNameToMatch('learner...');
+    subscriptionDashboardPage.expectSubscriptionLastNameToMatch('learner...');
+    users.logout();
+
+    users.login('creator2@subscriptions.com');
+    creatorDashboardPage.get();
+    browser.waitForAngular();
+    creatorDashboardPage.navigateToSubscriptionDashboard();
+    subscriptionDashboardPage.expectSubscriptionCountToEqual(1);
+    subscriptionDashboardPage.expectSubscriptionLastNameToMatch('learner...');
     users.logout();
   });
